% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l4testing.R
\name{crossUStatL4}
\alias{crossUStatL4}
\title{Title}
\usage{
crossUStatL4(X, m)
}
\arguments{
\item{X}{\eqn{n \times p} matrix of data.}

\item{m}{Cross threshold value. Should be between 3 and \eqn{n}.}
}
\value{
The cross U-Statistic for the \eqn{\ell_4}-norm calculated as \deqn{
\widetilde \theta_{(m+1):k} = \sum_{1 \leq i_1,...,i_{3} \leq m}^{*}\sum_{j = m+1}^k\left(\sum_{l = 1}^pX_{i_1,l}X_{i_2,l}X_{i_{3},l}X_{j,l}\right)
}
}
\description{
Title
}
\examples{
X <- matrix(rnorm(100 * 2), 100, 2)
crossUStatL4(X, 50)
}
