% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l4testing.R
\name{crossWStatL4}
\alias{crossWStatL4}
\title{The Self-Normalized Cross U-Statistic for \eqn{\ell_4} Norm Testing}
\usage{
crossWStatL4(data, m)
}
\arguments{
\item{data}{\eqn{n \times p} matrix of data.}

\item{m}{Cross threshold value. Should be between 3 and \eqn{n}.}
}
\value{
The self-normalized cross U-Statistic for the \eqn{\ell_4}-norm
The cross U-Statistic is given as \deqn{
\widetilde \theta_{(m+1):k} = \sum_{1 \leq i_1,...,i_{3} \leq m}^{*}\sum_{j = m+1}^k\left(\sum_{l = 1}^pX_{i_1,l}X_{i_2,l}X_{i_{3},l}X_{j,l}\right)
,} the self normalizer is given as \deqn{
V_{(m+1):n} = \frac{1}{n-m}\sum_{k = m+1}^n\left(\widetilde \theta_{(m+1):k} - \frac{k-m}{n-m}\widetilde \theta_{(m+1):n}\right)^2.
,} and the self normalized statistic is given as \deqn{
W := \frac{\widetilde \theta_{(m+1):n}^2}{V_{(m+1):n}}
}
}
\description{
The Self-Normalized Cross U-Statistic for \eqn{\ell_4} Norm Testing
}
\examples{
set.seed(12)
# Generate Data
data <- matrix(rnorm(100 * 2), 100, 2)
# Calculate U-Statistic
crossWStatL4(data, 50)
}
