% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{dcrossW}
\alias{dcrossW}
\title{Approximate Density of Cross W Distribution}
\usage{
dcrossW(x)
}
\arguments{
\item{x}{Vector of quantiles.}
}
\value{
Approximate kernel density estimate of the distribution based on \eqn{1e6} simulations of the random variable defined as
\deqn{
W:=\frac{{B(1)^2}}{\int_{0}^1(B(r) - rB(1))^2dr}
}
}
\description{
Approximate Density of Cross W Distribution
}
\examples{
require(graphics)

x<- seq(0, 1, length.out = 1e3)
q <- dcrossW(x)
plot(x, q, type = 'l')
}
