% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{pcrossW}
\alias{pcrossW}
\title{Empirical CDF of Cross W Statistic}
\usage{
pcrossW(q)
}
\arguments{
\item{q}{Vector of quantiles}
}
\value{
Empirical CDF of the distribution based on \eqn{1e6} simulations of the random variable defined as
\deqn{
W:=\frac{{B(1)^2}}{\int_{0}^1(B(r) - rB(1))^2dr}
}
}
\description{
Empirical CDF of Cross W Statistic
}
\examples{
require(graphics)

q <- seq(0, 200, length.out = 1e3)
p <- pcrossW(q)
plot(q, p, type = 'l')
}
