% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{qcrossW}
\alias{qcrossW}
\title{Approximate Cross W Distribution Quantiles}
\usage{
qcrossW(p)
}
\arguments{
\item{p}{Vector of probabilities}
}
\value{
Approximate quantile of the distribution based on \eqn{1e6} simulations of the random variable defined as
\deqn{
W:=\frac{{B(1)^2}}{\int_{0}^1(B(r) - rB(1))^2dr}
}
}
\description{
Approximate Cross W Distribution Quantiles
}
\examples{
require(graphics)

p <- seq(0, 1, length.out = 1e3)
q <- qcrossW(p)
plot(p, q, type = 'l')
}
