% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{rcrossW}
\alias{rcrossW}
\title{Approximate random generation of Cross W random variables}
\usage{
rcrossW(n, p = 200)
}
\arguments{
\item{n}{Number of observations}

\item{p}{Number of independent random variables to use per observation. A larger number will give higher precision but larger computation times.}
}
\value{
A vector of random variables approximately from the distribution
\deqn{
W:=\frac{{B(1)^2}}{\int_{0}^1(B(r) - rB(1))^2dr}
}.
Random variables are generated using the KL expansion of the Brownian bridge.
\deqn{
W = \frac{Z_0}{\sqrt{\sum_{k=1}^{p} \frac{Z_k^2}{(\pi k)^2}}}
} where \eqn{Z_k} are i.i.d. standard normal random variables.
}
\description{
Approximate random generation of Cross W random variables
}
\examples{
require(graphics)

set.seed(1213530)
W <- rcrossW(1000)
plot(density(W))
}
