% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l4testing.R
\name{crossUStatL4}
\alias{crossUStatL4}
\title{The Cross U-Statistic for \eqn{\ell_4} Norm Testing}
\usage{
crossUStatL4(data, m)
}
\arguments{
\item{data}{\eqn{n \times p} matrix of data.}

\item{m}{Cross threshold value. Should be between 3 and \eqn{n}.}
}
\value{
The cross U-Statistic for the \eqn{\ell_4}-norm calculated as \deqn{
\widetilde \theta_{(m+1):k} = \sum_{1 \leq i_1,...,i_{3} \leq m}^{*}\sum_{j = m+1}^k\left(\sum_{l = 1}^pX_{i_1,l}X_{i_2,l}X_{i_{3},l}X_{j,l}\right)
}
}
\description{
The Cross U-Statistic for \eqn{\ell_4} Norm Testing
}
\examples{
set.seed(12)
# Generate Data
data <- matrix(rnorm(100 * 2), 100, 2)
# Calculate U-Statistic
crossUStatL4(data, 50)
}
